+++ CHANGELOG.md
@@ -35,6 +35,24 @@ Section Order:
 ### Security
 -->
 
+### Fixed
+
+- FC can't see their own fleet when not in selected group(s)
+
+### Changed
+
+- `views.ajax_fleet_details` refactored
+- `tasks._process_fleet` refactored
+- `tasks.open_fleet` refactored
+- `tasks.send_fleet_invitation` refactored
+- `tasks.check_fleet_adverts` refactored
+- `tasks.get_fleet_composition` refactored
+- Fleet creation logic refactored, and better error handling added
+
+### Removed
+
+- MOTD from fleet details form, as it cannot be formatted properly and might delete the ingame MOTD when not set
+
 ## [2.6.0] - 2025-08-14
 
 ### Changed